Rename controlsListDomTokenList.js to src/mapml/utils/DOMTokenList.js Clean up white space.
